refactored lenses
